+++ lib/agent0/agent0/interactive_fuzz/__init__.py
@@ -1,6 +1,3 @@
 """Shared functions for interactive fuzz testing."""
 
-from .close_random_trades import close_random_trades
-from .generate_trade_list import generate_trade_list
-from .open_random_trades import open_random_trades
-from .setup_fuzz import setup_fuzz
+from .helpers import close_random_trades, generate_trade_list, open_random_trades, setup_fuzz
