+++ lib/agent0/agent0/interactive_fuzz/fuzz_long_short_maturity_values.py
@@ -10,9 +10,7 @@ import numpy as np
 from fixedpointmath import FixedPoint
 
 from agent0.hyperdrive.interactive.event_types import OpenLong, OpenShort
-from agent0.interactive_fuzz.generate_trade_list import generate_trade_list
-from agent0.interactive_fuzz.open_random_trades import open_random_trades
-from agent0.interactive_fuzz.setup_fuzz import setup_fuzz
+from agent0.interactive_fuzz import generate_trade_list, open_random_trades, setup_fuzz
 
 # main script has a lot of stuff going on
 # pylint: disable=too-many-locals
