+++ lib/agent0/agent0/interactive_fuzz/helpers/generate_trade_list.py
