+++ lib/agent0/agent0/interactive_fuzz/helpers/open_random_trades.py
